% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_embeddings.R
\name{analyze_embeddings}
\alias{analyze_embeddings}
\title{Analyze Embeddings}
\usage{
analyze_embeddings(
  df,
  method = "centroid",
  python_path = NULL,
  gee_project = NULL,
  cv = FALSE
)
}
\arguments{
\item{df}{Data frame containing columns A00 to A63.}

\item{method}{Mapping method: 'centroid' or 'ridge'.}
}
\value{
A list containing the results.
}
\description{
This function calculates the similarity between each location's embedding and the species centroid (mean embedding).
}
\keyword{internal}
