% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_embeddings.R
\name{ee_auth_service}
\alias{ee_auth_service}
\title{Initialize Earth Engine with Service Account}
\usage{
ee_auth_service(
  json_path = Sys.getenv("GEE_SERVICE_ACCOUNT_KEY"),
  venv_path = NULL
)
}
\arguments{
\item{json_path}{Path to the service account JSON key. Defaults to GEE_SERVICE_ACCOUNT_KEY env var.}

\item{venv_path}{Path to the Python virtual environment.}
}
\description{
Initialize Earth Engine with Service Account
}
\keyword{internal}
