% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_autoSDM}
\alias{install_autoSDM}
\title{Install autoSDM Python Dependencies}
\usage{
install_autoSDM(method = "virtualenv", envname = "r-autoSDM")
}
\arguments{
\item{method}{Installation method. Pass "auto" to let reticulate decide.}

\item{envname}{The name, or full path, of the environment in which Python packages are to be installed.}
}
\description{
This function installs the required Python dependencies for autoSDM
using reticulate.
}
